Refactor test class `EventCatcher` into utils to make accessible to other tests
