+++ tests/functional/manifest_validations/test_check_for_spaces_in_model_names.py
@@ -1,21 +1,11 @@
 import pytest
 
-from dataclasses import dataclass, field
 from dbt.cli.main import dbtRunner
-from dbt_common.events.base_types import BaseEvent, EventLevel, EventMsg
+from dbt_common.events.base_types import EventLevel
 from dbt.events.types import SpacesInModelNameDeprecation, TotalModelNamesWithSpacesDeprecation
 from dbt.tests.util import update_config_file
-from typing import Dict, List
-
-
-@dataclass
-class EventCatcher:
-    event_to_catch: BaseEvent
-    caught_events: List[EventMsg] = field(default_factory=list)
-
-    def catch(self, event: EventMsg):
-        if event.info.name == self.event_to_catch.__name__:
-            self.caught_events.append(event)
+from tests.functional.utils import EventCatcher
+from typing import Dict
 
 
 class TestSpacesInModelNamesHappyPath:
