+++ tests/functional/utils.py
@@ -1,7 +1,9 @@
 import os
 from contextlib import contextmanager
+from dataclasses import dataclass, field
 from datetime import datetime
-from typing import Optional
+from dbt_common.events.base_types import BaseEvent, EventMsg
+from typing import List, Optional
 from pathlib import Path
 
 
@@ -17,3 +19,13 @@ def up_one(return_path: Optional[Path] = None):
 
 def is_aware(dt: datetime) -> bool:
     return dt.tzinfo is not None and dt.tzinfo.utcoffset(dt) is not None
+
+
+@dataclass
+class EventCatcher:
+    event_to_catch: BaseEvent
+    caught_events: List[EventMsg] = field(default_factory=list)
+
+    def catch(self, event: EventMsg):
+        if event.info.name == self.event_to_catch.__name__:
+            self.caught_events.append(event)
